Access control with user relation done

+++ resources/views/posts/show.blade.php
@@ -10,20 +10,29 @@
     </p>
     </div>
 
-    <small class="text-warning bg-dark">Written on {{$post->created_at}}</small>
+    <small class="text-warning bg-dark">Written on {{$post->created_at}} by {{$post->user->name}}</small>
 
     <div>
         <hr>
-        <a href="/lsapp/public/posts/{{$post->id}}/edit" class="btn btn-primary">Edit</a>
-    <a href="../posts" class="btn btn-warning">Go Back</a>
-    </div>
-    <hr>
-    
-    {!! Form::open(['action' => ['PostsController@destroy', $post->id], 'method' => 'DELETE','class' =>'pull-right']) !!}
+        <a href="../posts" class="btn btn-warning">Go Back</a>
+            @if (!Auth::guest())
+                @if (Auth::user()->id== $post->user_id)
+            
+               
+            
+      
+                        <a href="/lsapp/public/posts/{{$post->id}}/edit" class="btn btn-primary">Edit</a>
 
-    
-    {{Form::submit('Delete',['class' =>'btn btn-danger'])}}
-    {!!Form::close()!!}
+                    </div>
+                    <hr>
+                    
+                    {!! Form::open(['action' => ['PostsController@destroy', $post->id], 'method' => 'DELETE','class' =>'pull-right']) !!}
+
+                    
+                    {{Form::submit('Delete',['class' =>'btn btn-danger'])}}
+                    {!!Form::close()!!}
+                 @endif
+    @endif
 
 </div>
 @endsection
